requirements.txt
flask
flask-cors
numpy
faiss-cpu
PyMuPDF
sentence-transformers
google-generativeai
